+++ opengb/printer/dummy.py
@@ -86,6 +86,11 @@ class Dummy(IPrinter):
                                            'head|{0}, length|{1}, rate|'
                                            '{2}'.format(head, length, rate))
 
+    def set_extrude_override(self, percent):
+        self._callbacks.log(logging.DEBUG, 'Extrude override set to '
+                                           '{0} percent'.format(percent))
+        self._callbacks.extrude_override_change(percent)
+
     def enable_steppers(self):
         self._callbacks.log(logging.DEBUG, 'Steppers enabled')
         self._callbacks.steppers_update(True)