+++ opengb/tests/test_server.py
@@ -5,6 +5,7 @@ TODO:
     
     * Add print-related API method tests
     * Add retract/unretract tests
+    * Add extrude override tests
 """
 
 import os